메인페이지 이동 및 변경
